refactor(ios): split BudgetDetailsView, cache sortedBudgets (PUL-209)
Extract sticky pager layer, FAB and skeleton into dedicated subviews.
Move scroll-fade opacity to an Observable tracker owned by the parent
but read only by the pager subview, so scroll ticks no longer re-eval
the parent body or its filter pipeline. Cache sorted month list in the
view model to drop the per-access O(n log n) sort and dedupe with
updateAdjacentBudgets. Adds Blur.maxRadiusStrong token; removes
swiftlint:disable file_length on the view (606 -> 392 lines).
